test(vats): refactor for reuse
