+++ packages/vats/test/test-boot.js
@@ -1,130 +1,26 @@
 // @ts-check
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
-import { E, Far, passStyleOf } from '@endo/far';
-import bundleSource from '@endo/bundle-source';
 import {
   makeFakeVatAdmin,
   zcfBundleCap,
 } from '@agoric/zoe/tools/fakeVatAdmin.js';
-import buildManualTimer from '@agoric/zoe/tools/manualTimer.js';
+import bundleSource from '@endo/bundle-source';
+import { E, Far, passStyleOf } from '@endo/far';
 
 import { makeZoeKit } from '@agoric/zoe';
-import { buildRootObject } from '../src/core/boot.js';
 import { buildRootObject as buildPSMRootObject } from '../src/core/boot-psm.js';
+import { buildRootObject } from '../src/core/boot.js';
 import { bridgeCoreEval } from '../src/core/chain-behaviors.js';
 import { makePromiseSpace } from '../src/core/utils.js';
-import { buildRootObject as bankRoot } from '../src/vat-bank.js';
 import { buildRootObject as boardRoot } from '../src/vat-board.js';
-import { buildRootObject as ibcRoot } from '../src/vat-ibc.js';
-import { buildRootObject as mintsRoot } from '../src/vat-mints.js';
-import { buildRootObject as networkRoot } from '../src/vat-network.js';
-import { buildRootObject as priceAuthorityRoot } from '../src/vat-priceAuthority.js';
-import { buildRootObject as provisioningRoot } from '../src/vat-provisioning.js';
-import { buildRootObject as zoeRoot } from '../src/vat-zoe.js';
 
-import { devices } from './devices.js';
-
-const vatRoots = {
-  bank: bankRoot,
-  board: boardRoot,
-  ibc: ibcRoot,
-  mints: mintsRoot,
-  network: networkRoot,
-  priceAuthority: priceAuthorityRoot,
-  provisioning: provisioningRoot,
-  zoe: zoeRoot,
-};
-
-const noop = () => {};
-
-const makeMock = t =>
-  harden({
-    devices: {
-      command: /** @type { any } */ ({ registerInboundHandler: noop }),
-      mailbox: /** @type { any } */ ({
-        registerInboundHandler: noop,
-      }),
-      timer: /** @type { any } */ ({}),
-      plugin: /** @type { any } */ ({ registerReceiver: noop }),
-      ...devices,
-    },
-    vats: {
-      vattp: /** @type { any } */ (
-        Far('vattp', {
-          registerMailboxDevice: noop,
-          addRemote: () => harden({}),
-        })
-      ),
-      comms: Far('comms', {
-        addRemote: noop,
-        addEgress: noop,
-        addIngress: async () =>
-          harden({
-            getConfiguration: () => harden({ _: 'client configuration' }),
-          }),
-      }),
-      http: { setPresences: noop, setCommandDevice: noop },
-      spawner: {
-        buildSpawner: () => harden({ _: 'spawner' }),
-      },
-      timer: Far('TimerVat', {
-        createTimerService: async () => buildManualTimer(t.log),
-      }),
-      uploads: { getUploads: () => harden({ _: 'uploads' }) },
-
-      network: Far('network', {
-        registerProtocolHandler: noop,
-        bind: () => harden({ addListener: noop }),
-      }),
-    },
-  });
-
-const mockSwingsetVats = mock => {
-  const { admin: fakeVatAdmin } = makeFakeVatAdmin(() => {});
-  const fakeBundleCaps = new Map(); // {} -> name
-  const getNamedBundleCap = name => {
-    const bundleCap = harden({});
-    fakeBundleCaps.set(bundleCap, name);
-    return bundleCap;
-  };
-
-  const createVat = (bundleCap, options) => {
-    const name = fakeBundleCaps.get(bundleCap);
-    assert(name);
-    switch (name) {
-      case 'zcf':
-        return fakeVatAdmin.createVat(zcfBundleCap, options);
-      default: {
-        const buildRoot = vatRoots[name];
-        if (!buildRoot) {
-          throw Error(`TODO: load vat ${name}`);
-        }
-        const vatParameters = { ...options?.vatParameters };
-        if (name === 'zoe') {
-          // basic-behaviors.js:buildZoe() provides hard-coded zcf BundleName
-          // and vat-zoe.js ignores vatParameters, but this would be the
-          // preferred way to pass the name.
-          vatParameters.zcfBundleName = 'zcf';
-        }
-        return { root: buildRoot({}, vatParameters), admin: {} };
-      }
-    }
-  };
-  const createVatByName = name => {
-    const bundleCap = getNamedBundleCap(name);
-    return createVat(bundleCap);
-  };
-
-  const vats = {
-    ...mock.vats,
-    vatAdmin: /** @type { any } */ ({
-      createVatAdminService: () =>
-        Far('vatAdminSvc', { getNamedBundleCap, createVat, createVatByName }),
-    }),
-  };
-  return vats;
-};
+import {
+  makeMock,
+  mockDProxy,
+  mockPsmBootstrapArgs,
+  vatRoots,
+} from '../tools/boot-test-utils.js';
 
 const argvByRole = {
   chain: {
@@ -143,11 +39,14 @@ const argvByRole = {
 };
 const testRole = (ROLE, governanceActions) => {
   test(`test manifest permits: ${ROLE} gov: ${governanceActions}`, async t => {
-    const mock = makeMock(t);
+    const mock = makeMock(t.log);
     const root = buildRootObject(
-      // @ts-expect-error Device<T> is a little goofy
-      { D: d => d, logger: t.log },
-      { argv: argvByRole[ROLE], governanceActions },
+      { D: mockDProxy, logger: t.log },
+      {
+        argv: argvByRole[ROLE],
+        // @ts-expect-error XXX
+        governanceActions,
+      },
     );
 
     const fakeVatAdmin = makeFakeVatAdmin(() => {}).admin;
@@ -266,9 +165,12 @@ test('evaluateInstallation is available to core eval', async t => {
 
 test('bootstrap provides a way to pass items to CORE_EVAL', async t => {
   const root = buildRootObject(
-    // @ts-expect-error Device<T> is a little goofy
-    { D: d => d, logger: t.log },
-    { argv: argvByRole.chain, governanceActions: false },
+    { D: mockDProxy, logger: t.log },
+    {
+      argv: argvByRole.chain,
+      // @ts-expect-error XXX
+      governanceActions: false,
+    },
   );
 
   await E(root).produceItem('swissArmyKnife', [1, 2, 3]);
@@ -285,16 +187,10 @@ const psmParams = {
 };
 
 test(`PSM-only bootstrap`, async t => {
-  const mock = makeMock(t);
-  const root = buildPSMRootObject(
-    // @ts-expect-error Device<T> is a little goofy
-    { D: d => d, logger: t.log },
-    psmParams,
-  );
+  const root = buildPSMRootObject({ D: mockDProxy, logger: t.log }, psmParams);
 
-  const vats = mockSwingsetVats(mock);
-  const actual = await E(root).bootstrap(vats, mock.devices);
-  t.deepEqual(actual, undefined);
+  const returnVal = await E(root).bootstrap(...mockPsmBootstrapArgs(t.log));
+  t.is(returnVal, undefined);
 
   const agoricNames = /** @type {Promise<NameHub>} */ (
     E(root).consumeItem('agoricNames')
@@ -304,11 +200,7 @@ test(`PSM-only bootstrap`, async t => {
 });
 
 test('PSM-only bootstrap provides a way to pass items to CORE_EVAL', async t => {
-  const root = buildPSMRootObject(
-    // @ts-expect-error Device<T> is a little goofy
-    { D: d => d, logger: t.log },
-    psmParams,
-  );
+  const root = buildPSMRootObject({ D: mockDProxy, logger: t.log }, psmParams);
 
   await E(root).produceItem('swissArmyKnife', [1, 2, 3]);
   t.deepEqual(await E(root).consumeItem('swissArmyKnife'), [1, 2, 3]);
